+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed delegations not loaded on validator lists @faboweb
 - Reconnected event was not triggered after reconnection @faboweb
+- [\#2094](https://github.com/cosmos/voyager/pull/2094) fix Toolbar @faboweb
 
 ## [1.0.0-beta.7] - 2019-02-26